Move HVG subsetting function to utils
