+++ src/integrate/scvi/config.vsh.yaml
@@ -103,9 +103,11 @@ functionality:
   resources:
     - type: python_script
       path: script.py
+    - path: ../../utils.py
   test_resources:
     - type: python_script
       path: test.py
+    - path: ../../utils.py
     - path: ../../../resources_test/pbmc_1k_protein_v3/pbmc_1k_protein_v3_mms.h5mu
 platforms:
   - type: docker
