+++ src/integrate/scvi/script.py
@@ -1,6 +1,7 @@
 import mudata
 import scvi
 from torch.cuda import is_available as cuda_is_available
+from utils import subset_hvg
 try:
     from torch.backends.mps import is_available as mps_is_available
 except ModuleNotFoundError:
@@ -30,24 +31,6 @@ par = {
     "output_compression": "gzip",
     }
 ### VIASH END
-
-
-def subset_hvg(adata, hvg_col):
-    """Subset highly variable genes from AnnData object
-    
-    Parameters
-    ----------
-    adata : AnnData
-        Annotated data object
-    hvg_col : str
-        Name of the boolean column in `adata.var` that contains the information if genes should be used or not
-
-    Returns
-    -------
-    AnnData
-        Copy of `adata` with subsetted genes
-    """
-    return adata[:, adata.var[hvg_col]].copy()
 
 
 def main():
